backend/requirements.txt
fastapi
uvicorn[standard]
websockets
gradium
google-genai
httpx
python-dotenv
pytest
pytest-asyncio
